Finaliezed the package creation
