--- Code/ChristerpherHunter/Flask/lab_02/pw_genny.py
@@ -1,10 +1,0 @@
-
-
-@app.route("/")
-def index():
-
-    return render_template("index.html")
-
-
-if __name__ == "__main__":
-    app.run(debug=True)
