+++ Code/ChristerpherHunter/Flask/lab_02/pw_genny/__init__.py
@@ -1,5 +1,4 @@
-from models import Data
-from flask import Flask, render_template, request, redirect, url_for
+from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 
 
@@ -8,3 +7,5 @@ app = Flask(__name__)
 app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///site.db"
 
 db = SQLAlchemy(app)
+
+from pw_genny import routes
