+++ Code/ChristerpherHunter/Flask/lab_02/run.py
@@ -1,0 +1,4 @@
+from pw_genny import app
+
+if __name__ == "__main__":
+    app.run(debug=True)
